refactoring of test utilities
